+++ src/db/repositories/BaseRepository.js
@@ -1,0 +1,9 @@
+const dbManager = require('../DBManager');
+
+class BaseRepository {
+  get db() {
+    return dbManager.getAdapter();
+  }
+}
+
+module.exports = BaseRepository;